refactor(cli): group plugins/plugin-marketplace/skills into lib/plugins (Move 3, RUSH-2708)
Move the 3 flat plugin-domain files into apps/cli/src/lib/plugins/, depth-fix
their own imports, and repoint every consumer. Behavior-preserving; build clean.
(Teammate did the move; finalized by orchestrator.)
