+++ apps/cli/src/lib/__tests__/bugfixes.test.ts
@@ -4,7 +4,7 @@ import * as path from 'path';
 import * as os from 'os';
 import { symlinkAllowedDirs } from '../sandbox.js';
 import { validateJob, parseAtTime, writeJob, readJob, deleteJob, resolveJobPrompt } from '../routines.js';
-import { cleanOrphanedPluginSkills } from '../plugins.js';
+import { cleanOrphanedPluginSkills } from '../plugins/plugins.js';
 
 describe('Bug Fix: Path traversal in sandbox.ts', () => {
   let overlayHome: string;
