+++ apps/cli/src/lib/drift.ts
@@ -13,7 +13,7 @@ import { getGlobalDefault, listInstalledVersions } from './versions.js';
 import { loadManifest, isStale } from './staleness/index.js';
 import { diffVersionResources, type VersionResourceReport, type SourceLayerBehind } from './doctor-diff.js';
 import { diffVersionCommands, iterCommandsCapableVersions } from './commands.js';
-import { diffVersionSkills, iterSkillsCapableVersions } from './skills.js';
+import { diffVersionSkills, iterSkillsCapableVersions } from './plugins/skills.js';
 import { iterHooksCapableVersions, listUnmanagedHooksInVersionHome, checkVersionHookWiring } from './hooks.js';
 import { commitsBehindUpstream } from './git.js';
 import { getUserAgentsDir, getSystemAgentsDir, getEnabledExtraRepos } from './state.js';
